djbackendps/requirements.txt
django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
streamlit>=1.10.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
streamlit-option-menu>=0.3.0
requests>=2.25.0
